feature: refactor kubernetes resource package
