+++ scaleph-engine/scaleph-engine-flink-kubernetes/src/main/java/cn/sliew/scaleph/engine/flink/kubernetes/resource/definition/deployment/FlinkDeployment.java
@@ -16,8 +16,9 @@
  * limitations under the License.
  */
 
-package cn.sliew.scaleph.engine.flink.kubernetes.resource.template;
+package cn.sliew.scaleph.engine.flink.kubernetes.resource.definition.deployment;
 
+import cn.sliew.scaleph.engine.flink.kubernetes.operator.spec.FlinkDeploymentSpec;
 import cn.sliew.scaleph.engine.flink.kubernetes.operator.status.FlinkDeploymentStatus;
 import cn.sliew.scaleph.kubernetes.Constant;
 import cn.sliew.scaleph.kubernetes.resource.Resource;
@@ -32,12 +33,12 @@ import lombok.EqualsAndHashCode;
 @Group(Constant.GROUP)
 @Version(Constant.VERSION)
 @EqualsAndHashCode
-@JsonPropertyOrder({"apiVersion", "kind", "metadata", "spec", "status"})
-public class FlinkTemplate extends CustomResource<FlinkTemplateSpec, FlinkDeploymentStatus> implements Resource {
+@JsonPropertyOrder({"apiVersion", "kind", "metadata", "spec"})
+public class FlinkDeployment extends CustomResource<FlinkDeploymentSpec, FlinkDeploymentStatus> implements Resource {
 
     @Override
     public String getKind() {
-        return Constant.FLINK_TEMPLATE;
+        return Constant.FLINK_DEPLOYMENT;
     }
 
     @Override
