+++ scaleph-engine/scaleph-engine-flink-kubernetes/src/main/java/cn/sliew/scaleph/engine/flink/kubernetes/resource/definition/deployment/FlinkDeploymentConverter.java
@@ -16,14 +16,13 @@
  * limitations under the License.
  */
 
-package cn.sliew.scaleph.engine.flink.kubernetes.resource.sessioncluster;
+package cn.sliew.scaleph.engine.flink.kubernetes.resource.definition.deployment;
 
-import cn.sliew.scaleph.engine.flink.kubernetes.operator.spec.FlinkSessionClusterSpec;
+import cn.sliew.scaleph.config.resource.ResourceLabels;
+import cn.sliew.scaleph.engine.flink.kubernetes.operator.spec.FlinkDeploymentSpec;
 import cn.sliew.scaleph.engine.flink.kubernetes.operator.spec.FlinkVersion;
-import cn.sliew.scaleph.engine.flink.kubernetes.operator.spec.KubernetesDeploymentMode;
-import cn.sliew.scaleph.engine.flink.kubernetes.service.dto.WsFlinkKubernetesSessionClusterDTO;
+import cn.sliew.scaleph.engine.flink.kubernetes.service.dto.WsFlinkKubernetesDeploymentDTO;
 import cn.sliew.scaleph.engine.flink.kubernetes.service.vo.KubernetesOptionsVO;
-import cn.sliew.scaleph.kubernetes.Constant;
 import cn.sliew.scaleph.kubernetes.resource.ResourceConverter;
 import io.fabric8.kubernetes.api.model.ObjectMetaBuilder;
 import org.apache.commons.lang3.EnumUtils;
@@ -31,19 +30,19 @@ import org.springframework.util.StringUtils;
 
 import java.util.Map;
 
-public enum FlinkSessionClusterConverter implements ResourceConverter<WsFlinkKubernetesSessionClusterDTO, FlinkSessionCluster> {
+public enum FlinkDeploymentConverter implements ResourceConverter<WsFlinkKubernetesDeploymentDTO, FlinkDeployment> {
     INSTANCE;
 
     @Override
-    public FlinkSessionCluster convertTo(WsFlinkKubernetesSessionClusterDTO source) {
-        FlinkSessionCluster sessionCluster = new FlinkSessionCluster();
+    public FlinkDeployment convertTo(WsFlinkKubernetesDeploymentDTO source) {
+        FlinkDeployment deployment = new FlinkDeployment();
         ObjectMetaBuilder builder = new ObjectMetaBuilder(true);
-        String name = StringUtils.hasText(source.getSessionClusterId()) ? source.getSessionClusterId() : source.getName();
+        String name = StringUtils.hasText(source.getDeploymentId()) ? source.getDeploymentId() : source.getName();
         builder.withName(name);
         builder.withNamespace(source.getNamespace());
-        builder.withLabels(Map.of(Constant.SCALEPH_NAME, source.getName()));
-        sessionCluster.setMetadata(builder.build());
-        FlinkSessionClusterSpec spec = new FlinkSessionClusterSpec();
+        builder.withLabels(Map.of(ResourceLabels.SCALEPH_LABEL_NAME, source.getName()));
+        deployment.setMetadata(builder.build());
+        FlinkDeploymentSpec spec = new FlinkDeploymentSpec();
         KubernetesOptionsVO kuberenetesOptions = source.getKubernetesOptions();
         if (kuberenetesOptions != null) {
             spec.setImage(kuberenetesOptions.getImage());
@@ -51,43 +50,41 @@ public enum FlinkSessionClusterConverter implements ResourceConverter<WsFlinkKub
             spec.setServiceAccount(kuberenetesOptions.getServiceAccount());
             spec.setFlinkVersion(EnumUtils.getEnum(FlinkVersion.class, kuberenetesOptions.getFlinkVersion()));
         }
-        spec.setFlinkConfiguration(source.getFlinkConfiguration());
         spec.setJobManager(source.getJobManager());
         spec.setTaskManager(source.getTaskManager());
         spec.setPodTemplate(source.getPodTemplate());
+        spec.setFlinkConfiguration(source.getFlinkConfiguration());
+        spec.setLogConfiguration(source.getLogConfiguration());
         spec.setIngress(source.getIngress());
-        spec.setMode(KubernetesDeploymentMode.NATIVE);
-        sessionCluster.setSpec(spec);
-        return sessionCluster;
+        deployment.setSpec(spec);
+        return deployment;
     }
 
     @Override
-    public WsFlinkKubernetesSessionClusterDTO convertFrom(FlinkSessionCluster target) {
-        WsFlinkKubernetesSessionClusterDTO dto = new WsFlinkKubernetesSessionClusterDTO();
+    public WsFlinkKubernetesDeploymentDTO convertFrom(FlinkDeployment target) {
+        WsFlinkKubernetesDeploymentDTO dto = new WsFlinkKubernetesDeploymentDTO();
         String name = target.getMetadata().getName();
         if (target.getMetadata().getLabels() != null) {
             Map<String, String> labels = target.getMetadata().getLabels();
-            name = labels.computeIfAbsent(Constant.SCALEPH_NAME, key -> target.getMetadata().getName());
+            name = labels.computeIfAbsent(ResourceLabels.SCALEPH_LABEL_NAME, key -> target.getMetadata().getName());
         }
         dto.setName(name);
-        dto.setSessionClusterId(target.getMetadata().getName());
+        dto.setDeploymentId(target.getMetadata().getName());
         dto.setNamespace(target.getMetadata().getNamespace());
-        dto.setSessionClusterId(target.getMetadata().getUid());
-        FlinkSessionClusterSpec spec = target.getSpec();
-        KubernetesOptionsVO kuberenetesOptions = new KubernetesOptionsVO();
-        if (kuberenetesOptions != null) {
-            kuberenetesOptions.setImage(spec.getImage());
-            kuberenetesOptions.setImagePullPolicy(spec.getImagePullPolicy());
-            kuberenetesOptions.setServiceAccount(spec.getServiceAccount());
-            if (spec.getFlinkVersion() != null) {
-                kuberenetesOptions.setFlinkVersion(spec.getFlinkVersion().name());
-            }
+        FlinkDeploymentSpec spec = target.getSpec();
+        KubernetesOptionsVO optionsVO = new KubernetesOptionsVO();
+        optionsVO.setImage(spec.getImage());
+        optionsVO.setImagePullPolicy(spec.getImagePullPolicy());
+        optionsVO.setServiceAccount(spec.getServiceAccount());
+        if (spec.getFlinkVersion() != null) {
+            optionsVO.setFlinkVersion(spec.getFlinkVersion().name());
         }
-        dto.setKubernetesOptions(kuberenetesOptions);
-        dto.setFlinkConfiguration(spec.getFlinkConfiguration());
+        dto.setKubernetesOptions(optionsVO);
         dto.setJobManager(spec.getJobManager());
         dto.setTaskManager(spec.getTaskManager());
         dto.setPodTemplate(spec.getPodTemplate());
+        dto.setFlinkConfiguration(spec.getFlinkConfiguration());
+        dto.setLogConfiguration(spec.getLogConfiguration());
         dto.setIngress(spec.getIngress());
         return dto;
     }
