+++ scaleph-engine/scaleph-engine-flink-kubernetes/src/main/java/cn/sliew/scaleph/engine/flink/kubernetes/resource/definition/logging/Logging.java
@@ -16,14 +16,13 @@
  * limitations under the License.
  */
 
-package cn.sliew.scaleph.engine.flink.kubernetes.resource.savepoint;
+package cn.sliew.scaleph.engine.flink.kubernetes.resource.definition.logging;
 
 import lombok.Data;
 
 @Data
-public class RestoreStrategy {
+public class Logging {
 
-    private String type;
-    private String initialSavepointPath;
-    private Boolean allowNonRestoredState;
+    private String name;
+    private String level;
 }
