+++ scaleph-engine/scaleph-engine-flink-kubernetes/src/main/java/cn/sliew/scaleph/engine/flink/kubernetes/resource/handler/SqlScriptFactory.java
@@ -16,10 +16,10 @@
  * limitations under the License.
  */
 
-package cn.sliew.scaleph.engine.flink.kubernetes.resource.definition;
+package cn.sliew.scaleph.engine.flink.kubernetes.resource.handler;
 
 import cn.sliew.scaleph.config.resource.ResourceNames;
-import cn.sliew.scaleph.engine.flink.kubernetes.resource.job.FlinkDeploymentJob;
+import cn.sliew.scaleph.engine.flink.kubernetes.resource.definition.job.FlinkDeploymentJob;
 import cn.sliew.scaleph.engine.flink.kubernetes.service.dto.WsFlinkKubernetesJobDTO;
 import cn.sliew.scaleph.kubernetes.resource.custom.ResourceCustomizer;
 import io.fabric8.kubernetes.api.model.*;
