+++ scaleph-engine/scaleph-engine-flink-kubernetes/src/main/java/cn/sliew/scaleph/engine/flink/kubernetes/service/impl/WsFlinkKubernetesTemplateServiceImpl.java
@@ -27,8 +27,8 @@ import cn.sliew.scaleph.engine.flink.kubernetes.operator.spec.IngressSpec;
 import cn.sliew.scaleph.engine.flink.kubernetes.operator.spec.JobManagerSpec;
 import cn.sliew.scaleph.engine.flink.kubernetes.operator.spec.TaskManagerSpec;
 import cn.sliew.scaleph.engine.flink.kubernetes.operator.util.TemplateMerger;
-import cn.sliew.scaleph.engine.flink.kubernetes.resource.template.FlinkTemplate;
-import cn.sliew.scaleph.engine.flink.kubernetes.resource.template.FlinkTemplateConverter;
+import cn.sliew.scaleph.engine.flink.kubernetes.resource.definition.template.FlinkTemplate;
+import cn.sliew.scaleph.engine.flink.kubernetes.resource.definition.template.FlinkTemplateConverter;
 import cn.sliew.scaleph.engine.flink.kubernetes.service.WsFlinkKubernetesTemplateService;
 import cn.sliew.scaleph.engine.flink.kubernetes.service.convert.WsFlinkKubernetesTemplateConvert;
 import cn.sliew.scaleph.engine.flink.kubernetes.service.dto.WsFlinkKubernetesTemplateDTO;
